dtrace: fix lquantize for 32-bit overflow on values
Fix dtrace_aggregate_lquantize() so that it no longer truncates
value to or computes bin index in 32 bits.  Linux bug is
26268136 dtrace_aggregate_lquantize() suffers from 32-bit overflow
It references a corresponding Solaris bug.

Orabug: 26268136

+++ dtrace/dtrace_probe_ctx.c
@@ -166,7 +166,7 @@ void dtrace_aggregate_lquantize(uint64_t *lquanta, uint64_t nval,
 	int32_t		base = DTRACE_LQUANTIZE_BASE(arg);
 	uint16_t	step = DTRACE_LQUANTIZE_STEP(arg);
 	uint16_t	levels = DTRACE_LQUANTIZE_LEVELS(arg);
-	int32_t		val = (int32_t)nval, level;
+	int64_t		val = (int64_t)nval, level;
 
 	ASSERT(step != 0);
 	ASSERT(levels != 0);
